refactor: Move date formatting utilities to shared users utils
